secure_file_share/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==0.7.1
PyJWT==2.8.0
python-dotenv==1.0.1
cryptography==41.0.7
email-validator==2.0.0